Fix for exact id aura triggers
Note: This is not a clean fix, not all spellIds for a spellname are
supported. It is not clear what spellId our custom UnitAura shall return.
The way it is implemented now it always returns the HIGHEST spellId.
If one wants to implement it in a different way you would have to
rewrite Bufftrigger2.